Refactored main index.tsx file and organized code in ~/src/*
