+++ app/auth/login.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import { Alert, Text, TextInput, TouchableOpacity, View } from "react-native";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
 
-import { Button } from "~/components/Button";
+import { Button } from "~/src/features/components/Button";
 
 export default function LoginScreen() {
   const [utEid, setUtEid] = useState("");
