+++ app/auth/mobility-preferences.tsx
@@ -11,7 +11,7 @@ import {
 } from "react-native";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
 
-import { Button } from "~/components/Button";
+import { Button } from "~/src/features/components/Button";
 import { apiClient } from "~/utils/api-client";
 import { useAuth } from "~/utils/useAuth";
