+++ app/auth/signup.tsx
@@ -5,7 +5,7 @@ import {
 } from "react-native";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
 
-import { Button } from "~/components/Button";
+import { Button } from "~/src/features/components/Button";
 
 export default function SignupScreen() {
   const insets = useSafeAreaInsets();
