+++ app/auth/ut-eid-coming-soon.tsx
@@ -3,7 +3,7 @@ import { CaretLeft } from "phosphor-react-native";
 import { Text, TouchableOpacity, View } from "react-native";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
 
-import { Button } from "~/components/Button";
+import { Button } from "~/src/features/components/Button";
 
 export default function UTEidComingSoonScreen() {
   const insets = useSafeAreaInsets();
