+++ src/features/components/AvoidanceAreaDetails.tsx
