+++ src/features/components/POIBottomSheet.tsx
@@ -11,7 +11,7 @@ import { typography } from '~/utils/typography';
 import useMapIcons from "~/utils/useMapIcons";
 import { getCardinalLabel, getCardinalLabelFromNeighbors } from "~/utils/utils";
 
-import buildingsData from '../assets/geojson/buildings_simple.json';
+import buildingsData from '~/assets/geojson/buildings_simple.json';
 
 interface POIData {
   poi: any;
